[IMP] report_qweb_signer use absolute path for certificate files
