+++ report_qweb_signer/demo/report_certificate_demo.xml
@@ -8,8 +8,8 @@ License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
     <record id="demo_certificate_test" model="report.certificate">
         <field name="company_id" ref="base.main_company" />
         <field name="name">Test OCA certificate</field>
-        <field name="path">test.p12</field>
-        <field name="password_file">test.passwd</field>
+        <field name="path">.../static/certificate/test.p12</field>
+        <field name="password_file">.../static/certificate/test.passwd</field>
         <field name="model_id" ref="base.model_res_partner" />
         <field name="allow_only_one" eval="True" />
         <field
