+++ report_qweb_signer/tests/test_report_qweb_signer.py
@@ -1,7 +1,19 @@
 # Copyright 2017 Tecnativa - Pedro M. Baeza
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
+import os
+
 from odoo.tests.common import HttpCase
+
+
+def _normalize_filepath(path):
+    path = path or ""
+    path = path.strip()
+    if not os.path.isabs(path):
+        me = os.path.dirname(__file__)
+        path = "{}/../static/certificate/".format(me) + path
+    path = os.path.normpath(path)
+    return path if os.path.exists(path) else False
 
 
 class TestReportQwebSigner(HttpCase):
@@ -13,6 +25,12 @@ class TestReportQwebSigner(HttpCase):
         ).with_context(force_report_rendering=True)
 
     def test_report_qweb_signer(self):
+        self.report = self.env.ref("report_qweb_signer.demo_certificate_test").write(
+            {
+                "path": _normalize_filepath("test.p12"),
+                "password_file": _normalize_filepath("test.passwd"),
+            }
+        )
         self.report.render_qweb_pdf(self.partner.ids)
         # Reprint again for taking the PDF from attachment
         IrAttachment = self.env["ir.attachment"]
